+++ apps/mobile/src/components/episode/index.ts
@@ -1,4 +1,3 @@
 export * from './Episode';
 export * from './QualitySelect';
 export * from './OfflineEpisode';
-export { PlatformExplicit } from '../PlatformExplicit';
