+++ apps/mobile/src/components/episode/player/index.ts
@@ -1,1 +1,3 @@
 export * from './EpisodeImage';
+export * from './EpisodePlayer';
+export * from './EpisodeWatchProgress';
